test avrpool and mnist reader

+++ Neuro/ActivationFunctions/Sigmoid.cs
@@ -4,13 +4,13 @@ namespace Neuro.ActivationFunctions
 {
     public class Sigmoid : IActivationFunction
     {
-        public double Alpha { get; set; } = 1;
+        public double Alpha { get; set; } = 1d;
         public double MinRange { get; set; } = 0;
         public double MaxRange { get; set; } = 1;
         
         public double Activation(double x)
         {
-            return 1 / (1 + (double)Math.Exp(-Alpha * x));
+            return 1 / (1 + Math.Exp(-Alpha * x));
         }
 
         public double Derivative(double x)
